Combined async/sync functions, added some checking that int's are actually ints, added callback support to bufferSize
