+++ src/compress.cc
@@ -1,14 +1,111 @@
 #include "exports.h"
+#include "compress.h"
 
-void compressBufferFreeCallback(char *data, void *hint) {
-  tjFree((unsigned char*) data);
-}
+class CompressWorker : public Nan::AsyncWorker {
+  public:
+    CompressWorker(Nan::Callback *callback, unsigned char* srcData, uint32_t width, uint32_t stride, uint32_t height, uint32_t bpp, uint32_t format, uint32_t jpegSubsamp, int quality, v8::Local<v8::Object> &dstObject, unsigned char* dstData, uint32_t dstBufferLength)
+      : Nan::AsyncWorker(callback), srcData(srcData), width(width), stride(stride), height(height), bpp(bpp), format(format), jpegSubsamp(jpegSubsamp), quality(quality),
+        dstData(dstData), dstBufferLength(dstBufferLength) {
+        if (dstBufferLength > 0) {
+          SaveToPersistent("dstObject", dstObject);
+        }
+      }
+    ~CompressWorker() {}
 
-NAN_METHOD(CompressSync) {
-  int err;
-  const char* tjErr;
+    void Execute () {
+      int err;
 
-  if (info.Length() < 2) {
+      // Set up buffers if required
+      int flags = TJFLAG_FASTDCT;
+
+      this->dstLength = tjBufSize(this->width, this->height, this->jpegSubsamp);
+
+      if (this->dstBufferLength > 0) {
+        if (this->dstLength > this->dstBufferLength) {
+          SetErrorMessage("Pontentially insufficient output buffer");
+          return;
+        }
+        flags |= TJFLAG_NOREALLOC;
+      }
+
+      tjhandle handle = tjInitCompress();
+      if (handle == NULL) {
+        SetErrorMessage(tjGetErrorStr());
+        return;
+      }
+
+      err = tjCompress2(
+        handle,
+        this->srcData,
+        this->width,
+        this->stride * this->bpp,
+        this->height,
+        this->format,
+        &this->dstData,
+        &this->jpegSize,
+        this->jpegSubsamp,
+        this->quality,
+        flags
+      );
+
+      if (err != 0) {
+        SetErrorMessage(tjGetErrorStr());
+        tjDestroy(handle);
+        return;
+      }
+
+      err = tjDestroy(handle);
+      if (err != 0) {
+        SetErrorMessage(tjGetErrorStr());
+        if (this->dstBufferLength == 0) {
+          tjFree(dstData);
+        }
+        return;
+      }
+    }
+
+    void HandleOKCallback () {
+      Nan::HandleScope scope;
+
+      v8::Local<v8::Object> obj = Nan::New<v8::Object>();
+      v8::Local<v8::Object> dstObject;
+
+      if (this->dstBufferLength > 0) {
+        dstObject = GetFromPersistent("dstObject").As<v8::Object>();
+      }
+      else {
+        dstObject = Nan::NewBuffer((char*)this->dstData, this->jpegSize, compressBufferFreeCallback, NULL).ToLocalChecked();
+      }
+
+      obj->Set(Nan::New("data").ToLocalChecked(), dstObject);
+      obj->Set(Nan::New("size").ToLocalChecked(), Nan::New((uint32_t) this->jpegSize));
+
+      v8::Local<v8::Value> argv[] = {
+        Nan::Null(),
+        obj
+      };
+
+      callback->Call(2, argv);
+    }
+
+  private:
+    unsigned char* srcData;
+    uint32_t width;
+    uint32_t stride;
+    uint32_t height;
+    uint32_t bpp;
+    uint32_t format;
+    uint32_t jpegSubsamp;
+    int quality;
+    unsigned char* dstData;
+    uint32_t dstBufferLength;
+
+    uint32_t dstLength;
+    unsigned long jpegSize;
+};
+
+NAN_METHOD(Compress) {
+  if (info.Length() < 3) {
     return Nan::ThrowError(Nan::TypeError("Too few arguments"));
   }
 
@@ -24,7 +121,8 @@ NAN_METHOD(CompressSync) {
 
   // Output buffer
   v8::Local<v8::Object> dstObject;
-  bool havePreallocatedBuffer = false;
+  uint32_t dstBufferLength = 0;
+  unsigned char* dstData = NULL;
 
   // Options
   v8::Local<v8::Object> options = info[cursor++].As<v8::Object>();
@@ -32,16 +130,19 @@ NAN_METHOD(CompressSync) {
   if (node::Buffer::HasInstance(options) && info.Length() > cursor) {
     dstObject = options;
     options = info[cursor++].As<v8::Object>();
-    havePreallocatedBuffer = true;
+    dstBufferLength = node::Buffer::Length(dstObject);
+    dstData = (unsigned char*) node::Buffer::Data(dstObject);
   }
 
   if (!options->IsObject()) {
     return Nan::ThrowError(Nan::TypeError("Options must be an Object"));
   }
 
+  // Callback
+  Nan::Callback *callback = new Nan::Callback(info[cursor++].As<v8::Function>());
+
   // Format of input buffer
-  v8::Local<v8::Value> formatObject =
-    options->Get(Nan::New("format").ToLocalChecked());
+  v8::Local<v8::Value> formatObject = options->Get(Nan::New("format").ToLocalChecked());
 
   if (formatObject->IsUndefined()) {
     return Nan::ThrowError(Nan::TypeError("Missing format"));
@@ -52,49 +153,45 @@ NAN_METHOD(CompressSync) {
   // Figure out bpp from format (needed later for stride)
   uint32_t bpp;
   switch (format) {
-  case FORMAT_GRAY:
-    bpp = 1;
-    break;
-  case FORMAT_RGB:
-  case FORMAT_BGR:
-    bpp = 3;
-    break;
-  case FORMAT_RGBX:
-  case FORMAT_BGRX:
-  case FORMAT_XRGB:
-  case FORMAT_XBGR:
-  case FORMAT_RGBA:
-  case FORMAT_BGRA:
-  case FORMAT_ABGR:
-  case FORMAT_ARGB:
-    bpp = 4;
-    break;
-  default:
-    return Nan::ThrowError(Nan::TypeError("Invalid output format"));
+    case FORMAT_GRAY:
+      bpp = 1;
+      break;
+    case FORMAT_RGB:
+    case FORMAT_BGR:
+      bpp = 3;
+      break;
+    case FORMAT_RGBX:
+    case FORMAT_BGRX:
+    case FORMAT_XRGB:
+    case FORMAT_XBGR:
+    case FORMAT_RGBA:
+    case FORMAT_BGRA:
+    case FORMAT_ABGR:
+    case FORMAT_ARGB:
+      bpp = 4;
+      break;
+    default:
+      return Nan::ThrowError(Nan::TypeError("Invalid output format"));
   }
 
   // Subsampling
-  v8::Local<v8::Value> sampObject =
-    options->Get(Nan::New("subsampling").ToLocalChecked());
+  v8::Local<v8::Value> sampObject = options->Get(Nan::New("subsampling").ToLocalChecked());
 
-  uint32_t jpegSubsamp = sampObject->IsUndefined()
-    ? DEFAULT_SUBSAMPLING
-    : sampObject->Uint32Value();
+  uint32_t jpegSubsamp = sampObject->IsUndefined() ? DEFAULT_SUBSAMPLING : sampObject->Uint32Value();
 
   switch (jpegSubsamp) {
-  case SAMP_444:
-  case SAMP_422:
-  case SAMP_420:
-  case SAMP_GRAY:
-  case SAMP_440:
-    break;
-  default:
-    return Nan::ThrowError(Nan::TypeError("Invalid subsampling method"));
+    case SAMP_444:
+    case SAMP_422:
+    case SAMP_420:
+    case SAMP_GRAY:
+    case SAMP_440:
+      break;
+    default:
+      return Nan::ThrowError(Nan::TypeError("Invalid subsampling method"));
   }
 
   // Width
-  v8::Local<v8::Value> widthObject =
-    options->Get(Nan::New("width").ToLocalChecked());
+  v8::Local<v8::Value> widthObject = options->Get(Nan::New("width").ToLocalChecked());
 
   if (widthObject->IsUndefined()) {
     return Nan::ThrowError(Nan::TypeError("Missing width"));
@@ -103,8 +200,7 @@ NAN_METHOD(CompressSync) {
   uint32_t width = widthObject->Uint32Value();
 
   // Height
-  v8::Local<v8::Value> heightObject =
-    options->Get(Nan::New("height").ToLocalChecked());
+  v8::Local<v8::Value> heightObject = options->Get(Nan::New("height").ToLocalChecked());
 
   if (heightObject->IsUndefined()) {
     return Nan::ThrowError(Nan::TypeError("Missing height"));
@@ -113,75 +209,14 @@ NAN_METHOD(CompressSync) {
   uint32_t height = heightObject->Uint32Value();
 
   // Stride
-  v8::Local<v8::Value> strideObject =
-    options->Get(Nan::New("stride").ToLocalChecked());
+  v8::Local<v8::Value> strideObject = options->Get(Nan::New("stride").ToLocalChecked());
 
-  uint32_t stride = strideObject->IsUndefined()
-    ? width
-    : strideObject->Uint32Value();
+  uint32_t stride = strideObject->IsUndefined() ? width : strideObject->Uint32Value();
 
   // Quality
-  v8::Local<v8::Value> qualityObject =
-    options->Get(Nan::New("quality").ToLocalChecked());
+  v8::Local<v8::Value> qualityObject = options->Get(Nan::New("quality").ToLocalChecked());
 
-  int quality = qualityObject->IsUndefined()
-    ? DEFAULT_QUALITY
-    : qualityObject->Uint32Value();
+  int quality = qualityObject->IsUndefined() ? DEFAULT_QUALITY : qualityObject->Uint32Value();
 
-  // Set up buffers if required
-  int flags = TJFLAG_FASTDCT;
-  unsigned char* dstData = NULL;
-  uint32_t dstLength = tjBufSize(width, height, jpegSubsamp);
-
-  if (havePreallocatedBuffer) {
-    if (node::Buffer::Length(dstObject) < dstLength) {
-      return Nan::ThrowError("Potentially insufficient output buffer");
-    }
-
-    dstData = (unsigned char*) node::Buffer::Data(dstObject);
-    flags |= TJFLAG_NOREALLOC;
-  }
-
-  tjhandle handle = tjInitCompress();
-  if (handle == NULL) {
-    return Nan::ThrowError(tjGetErrorStr());
-  }
-
-  unsigned long jpegSize;
-  err = tjCompress2(
-    handle,
-    srcData,
-    width,
-    stride * bpp,
-    height,
-    format,
-    &dstData,
-    &jpegSize,
-    jpegSubsamp,
-    quality,
-    flags
-  );
-
-  if (err != 0) {
-    tjErr = tjGetErrorStr();
-    tjDestroy(handle);
-    return Nan::ThrowError(tjErr);
-  }
-
-  if (!havePreallocatedBuffer) {
-    dstObject = Nan::NewBuffer((char*) dstData, jpegSize,
-      compressBufferFreeCallback, NULL).ToLocalChecked();
-  }
-
-  err = tjDestroy(handle);
-  if (err != 0) {
-    return Nan::ThrowError(tjGetErrorStr());
-  }
-
-  // Unsure how to return a slice from here. Let's leave it to JS instead.
-  v8::Local<v8::Object> obj = Nan::New<v8::Object>();
-  obj->Set(Nan::New("data").ToLocalChecked(), dstObject);
-  obj->Set(Nan::New("size").ToLocalChecked(), Nan::New((uint32_t) jpegSize));
-
-  info.GetReturnValue().Set(obj);
+  Nan::AsyncQueueWorker(new CompressWorker(callback, srcData, width, stride, height,  bpp, format, jpegSubsamp, quality, dstObject, dstData, dstBufferLength));
 }
