--- src/compress.h
@@ -1,4 +1,0 @@
-#ifndef COMPRESS_H
-#define COMPRESS_H
-void compressBufferFreeCallback(char *data, void *hint);
-#endif
